Disable insertion of Interact buttons
Vocareum's URL generation doesn't support the kind of links we need
to use Berkeley's way of generating Interact buttons in the textbook
HTML.  Vocareum's links are not of the form PREFIX/notebookname.ipynb,
but instead have unique numbers.  Whereas Berkeley assumes that
prefix scheme.

This commit comments out the code to generate the buttons, so that
we'll still remember it in the future, and hopefully can re-instate
the buttons.

Until then we can grab Vocareum's links and put them as links on
our course website instead of embeded in the textbook itself.