Create output directory during notebook conversion.
The Python script that drives conversion of notebooks to HTML
assumed that the output directory would always exist, probably
because it always *did* exist for Berkeley, because they
had all the generated HTML files committed to the repo.
Removing them from the repo surfaced this bug.

This commit creates the directory, if needed, as part of conversion.