Fix MathJaX import
The Berkeley version of the book build file, book.json, imports the
MathJax plugin from Berkeley's own fork of that plugin.  There's no
need to do that AFAIK, and Berkeley even has a PR (containing a
bunch of other things) that seems to change the import to the
main version of the plugin.  Moreover, trying to build with the
import from Berkeley's fork currently produces an error.

This commit changes the import to the main MathJaX plugin, i.e.,
not Berkeley's fork of it.  The commit also adds a make target
to properly install the plugin.