Change deploy target to work at Cornell
The Berkeley deploy target publishes the book on gitbooks.
Cornell is instead hosting our own custom version of the book
on our course website. So the Berkeley deploy needs to be changed.

This commit makes that change.

Deployment requires write permission to the course website,
though not to the website repo.