Rewrite title page for Cornell Edition of book
The Berkeley Edition of the book links directly to its course website
and online version of the book.

This commit changes links and text on the title page and the side bar to
point to the Cornell course.  It also clearly identifies the text as the
Cornell Edition, gives credit to Clarkson and Udell for adaptations, and
states that we have permission to do so.

The link to the Berkeley Edition is left on the title page for sake of
scholarship, but buried further down to minimize chances for confusion.