Remove build dependency from deploy target
The makefile deploy target had the build target as a dependency.
That caused all the notebooks to be reconverted to HTML every
time the book was deployed.  At best that's unnecessary, and
at worst it deploys changes the author has not yet reviewed.
This commit removes the dependency, and reminds the author
that it's up to them to put their eyeballs on any edits
before deploying those edits.