Remove `read_url` from Chapter 1
The textbook used a function `real_url` to read the text of Huck Finn
and Little Women from a URL on the Berkeley textbook website.  Vocareum,
though, doesn't allow outgoing IP connections, so that code wouldn't
run for students in the Cornell course who were viewing the interactive
notebook in Vocareum's cloud.

This commit adds the data files to the textbook and changes the
`read_url` to a `read_file` function.