Remove dead notebook: Books.ipynb
The Books.ipynb notebook was in the repo but was no longer being
used as part of the book.  Instead, its contents had apparently
been refactored across three other notebooks that comprise Chapter
1.  It also contained code that would need to be fixed if it
were to stay in the Cornell version of the textbook, but there's
no reason to maintain it if it's dead.

This commit therefore removes the notebook.