Remove generated image files
The Berkeley textbook repo had a folder, notebooks-images/, of .png
files that are generated as part of the process of building the book.
There's no need for Cornell to have those generated files commited
to the repo.

This commit removes the files, adds the folder to the gitignore,
and adds code to the generation script to create the folder
if it doesn't already exist.