Remove generated directories from .gitignore
The directories notebooks-html and notebooks-images are produced
as output from conversion of .ipynb notebooks to HTML.  In principle
those directories shouldn't be checked in to source control.  But
unfortunately Gitbook is designed to exclude all files in .gitignore
from the book output.  So if we exclude the generated files from
the repo, they also will be excluded from the book.  There is
an open Issue for this:
  https://github.com/GitbookIO/gitbook/issues/931

For now we will have to live with not ignoring the directories.
So this commit records that fact in the .gitignore, and adds
a `make clean` target to remove the generated files, which is
useful before checking `git status`.