Enable Interact buttons, but not per-notebook
Vocareum currently supports linking only to a single URL for an
assignment-part, not to a notebook within that part.  So these
interact buttons simply link to the current assignment-part
in Vocareum for 2018sp that will contain the entire textbook.
Later I hope we can do direct links to notebooks, or perhaps
link to a separate part for each chapter.