Fix MathJaX plugin
The MathJaX plugin was not working in the Cornell version of the
textbook.  The reasons why are complicated.  This commit fixes it
by creating our own fork of that plugin and linking against it.

In more detail:
* The main plugin v2.0.0 at https://github.com/GitbookIO/plugin-mathjax/
  is actually broken, as can be seen in PR32.  It uses the wrong
  URL for MathJaX, which transitioned to a different CDN back
  in April 2017.
* The Berkeley textbook had fixed this by creating its own fork
  of the plugin based on v1.1.0 and updating the URL.  But their fork
  will not compile on OS X under the current version of Node.js.

I tried creating our own fork of the main plugin, v2.0.0, and
updating the URL.  Unfortunately, that won't compile against
Gitbook v3.x, which is what the textbook uses:  it would instead
need 4.0.0.alpha, which is still under development.

I discovered that the main plugin v1.1.2 does still compile
against the textbook under the current version of Node.js and
using Gitbook v3.x.  So I created a fork and branch of that
version and updated the URL.  That works!

This commit points the book to use our fork and that branch.